解决了 SQL 的一些 BUG

+++ sql_executor/parser/ast.h
@@ -152,7 +152,7 @@ namespace ast {
 
     struct FloatLit : public Value{
         float m_val;
-        FloatLit(int val) : m_val(val) {}
+        FloatLit(float val) : m_val(val) {}
     };
 
     struct StringLit : public Value{
